unidades funcionais implementadas e doctorId refatorado para doctor
