--- src/main/resources/application-dev.properties
@@ -1,11 +1,0 @@
-spring.datasource.url=jdbc:postgresql://localhost:5433/pacientesdb
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-spring.datasource.driver-class-name=org.postgresql.Driver
-
-spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-
-server.port=8083
